requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart